requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
requests>=2.31.0
cryptography>=41.0.0
Pillow>=10.0.0
